// app/models/client.model.ts
export class Client {
    klijentId?: number = 0;
    naziv?: string = '';
    adresa?: string = '';
    mjesto?: string = '';
    pdvbroj?: string = '';
    idbroj?: string = '';
    telefon?: string = '';
    image?: string = '';
    odgovornaOsoba?: string = '';
    email?: string = '';
    brojBankovnogRacuna?: string = '';
    isPotvrdjen?: boolean = false;
    potvrdjenMail?: boolean = false;
}